+++ CHANGELOG.md
@@ -16,6 +16,7 @@ minor release, the component will be purged, so be prepared (see `Updating` sect
 - New storage component for small objects named Peapod (#2453)
 - New `blobovnicza-to-peapod` tool providing blobovnicza-to-peapod data migration (#2453)
 - SN's version and capacity is announced via the attributes automatically but can be overwritten explicitly (#2455, #602)
+- `peapod` command for `neofs-lens` (#2507)
 
 ### Fixed
 - `neo-go` RPC connection loss handling (#1337)